refactor!: Move validate_schema to workspace.validators
